fix: restructure vitest using colocate approach
