+++ src/utils/actions.test.ts
@@ -1,0 +1,9 @@
+import { describe, it, expect } from 'vitest';
+// import { functionToTest } from './actions'; // Import functions as needed
+
+describe('Actions Utils (src/utils/actions.ts)', () => {
+  it('should have tests', () => {
+    // TODO: Add tests for actions utility functions
+    expect(true).toBe(true); // Placeholder assertion
+  });
+});
