+++ src/utils/email-templates.test.ts
@@ -1,0 +1,9 @@
+import { describe, it, expect } from 'vitest';
+// import { functionToTest } from './email-templates'; // Import functions as needed
+
+describe('Email Templates Utils (src/utils/email-templates.ts)', () => {
+  it('should have tests', () => {
+    // TODO: Add tests for email-templates utility functions
+    expect(true).toBe(true); // Placeholder assertion
+  });
+});
