+++ src/utils/user.test.ts
@@ -1,0 +1,9 @@
+import { describe, it, expect } from 'vitest';
+// import { functionToTest } from './user'; // Import functions as needed
+
+describe('User Utils (src/utils/user.ts)', () => {
+  it('should have tests', () => {
+    // TODO: Add tests for user utility functions
+    expect(true).toBe(true); // Placeholder assertion
+  });
+});
